added calc for total order cost

+++ DisneyFilmStore.Data/Film.cs
@@ -20,8 +20,8 @@ namespace DisneyFilmStore.Data
         [Required]
         public DateTime YearReleased { get; set; }
         [Required]
-        public double MemberCost { get; set; }
+        public decimal MemberCost { get; set; }
         [Required]
-        public double NonMemberCost { get; set; }
+        public decimal NonMemberCost { get; set; }
     }
 }
